requirements.txt
fastapi
uvicorn
SQLAlchemy
psycopg2-binary
alembic
python-jose[cryptography]
passlib[bcrypt]
pydantic
pydantic-settings
pytest
httpx
requests
python-multipart
pydantic[email]
python-dotenv
email-validator
fastapi_limiter
websockets
aiohttp
google-generativeai
bcrypt==4.1.2
authlib
sendgrid
Pillow
aiofiles
